python/pyproject.toml
[tool.poetry]
name = "pygame-samples-1"
version = "0.1.0"
description = ""
authors = ["Naohiro2g <587586+Naohiro2g@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pygame-ce = "^2.4.1"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
